Refactor to extract menu logic to a module
